Create Mastermind game
This is a game in which the player has to guess a four digit code within five tries. After each failed try a hint is provided based on player input, whether any of the given input is present in the 4-digit code and if yes, is it placed at the correct position.